feat: got quick add function to work with in memory

+++ models/src/budget_model.rs
@@ -5,12 +5,12 @@ pub struct Budget {
     pub id: u64,
     pub name: String,
     pub participants: Vec<String>,
-    pub balance: u64,
+    pub balance: f64,
     pub icon: u32,
 }
 
 impl Budget {
-    pub fn new(name: String, participants: Vec<String>, balance: u64, icon: u32) -> Budget {
+    pub fn new(name: String, participants: Vec<String>, balance: f64, icon: u32) -> Budget {
         Budget {
             id: 1,
             name,
